Update ESLint rules for react in JSX
Two rules were added to .eslintrc.cjs to improve the eslint configuration for our project. These rule updates turn off the requirement for React to be in scope when using JSX. This change is necessary because we moved to the new JSX transform introduced in React 17, which does not require React imports in files anymore.